fix bug

+++ src/components/Admin/UserRole/UserRoleDataTable.js
@@ -26,16 +26,22 @@ const UserRoleDataTable = () => {
   }
 
   const removeUserRole = async (userID, roleID) => {
-    await UserRoleService.deleteUserRole(userID, roleID)
-    .then(res => {
-        setUserRoleData();
-        alert("Delete Success");
-        // navigate("/user-role/read");
-    })
-    .catch(err => {
-      
-      alert("Error Ocurred in removeUserRole:" + err);
-    });
+
+    if(window.confirm("Delete item ?")){
+      await UserRoleService.deleteUserRole(userID, roleID)
+      .then(res => {
+          setUserRoleData();
+          alert("Delete Success");
+          // navigate("/user-role/read");
+      })
+      .catch(err => {
+        
+        alert("Error Ocurred in removeUserRole:" + err);
+      });
+    }
+    else{
+
+    }
   }
   
   useEffect(() => {
@@ -48,7 +54,7 @@ const UserRoleDataTable = () => {
   return (
     // Thêm className = "content-wrapper" vào tránh Navbar che chữ
     <div className="content-wrapper">
-      <div class="card-body" style={{textAlign:'center'}}>
+      <div className="card-body" style={{textAlign:'center'}}>
         <br>
         </br>
         <nav>
@@ -64,16 +70,15 @@ const UserRoleDataTable = () => {
         <div className="col-12">
           <h4 style={{textAlign:'center'}}>User Role List</h4>
 
-          <div class="container">
-            <div class="row">
-              <div class="col-12">
-                <table class="table table-bordered table-striped">
+          <div className="container">
+            <div className="row">
+              <div className="col-12">
+                <table className="table table-bordered table-striped">
                   <thead>
                     <tr>
                       {/* <th>ID</th> */}
                       <th>User</th>
                       <th>Role</th>
-                      <th scope="col">Action</th>
 
                     </tr>
                   </thead>
@@ -81,23 +86,25 @@ const UserRoleDataTable = () => {
 
                     {
                       userRole &&
-                      userRole.map((userRoles, index) => (
+                      userRole.map((userRoles,index) => (
 
-                        <tr>
+                        <tr key={index}>
                           {/* <th scope="row">{index}</th> */}
                           <td>{userRoles.user.username}</td>
                           <td>{userRoles.role.name}</td>
-                          <td style={{display:'inline-block'}}>
-                            <button>
-                                {/* useParam ở trang Edit sễ lấy 2 cái id  */}
-                                <Link to={"/admin/user-role/edit/" + userRoles.user.id + "/" + userRoles.role.id}>
-                                  Edit
-                                </Link>
+                          <td>
+                            <button className='btn btn-primary w-100' onClick={() => navigate("/admin/user-role/edit/" + userRoles.user.id + "/" + userRoles.role.id)}>
+                               
+                                  <i className="fa fa-edit"></i>
+                               
                             </button>
                             
-                            <button
-                              onClick={() => removeUserRole(userRoles.user.id, userRoles.role.id)} className="button"
-                            > Delete
+                          </td>
+                          <td>
+                            <button className='btn btn-primary  w-100'
+                              onClick={() => removeUserRole(userRoles.user.id, userRoles.role.id)}
+                            >                                   <i className="fa fa-trash"></i>
+
                             </button>
 
                           </td>
